requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
pycparser==2.23
python-decouple==3.8
python-telegram-bot==13.2
pytz==2025.2
requests==2.32.4
six==1.17.0
tornado==6.4.2
typing-extensions==4.13.2
tzdata==2025.3
tzlocal==5.2
urllib3==2.2.3